dl-api/requirements.txt
# Web Server & API Validation
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# Vector Database
chromadb>=0.4.24

# Deep Learning & Vision Pipeline
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0